[Apollo] Move checkpoint creation test
Move checkpoint creation test into a new checkpoints python module.
